Move Extension Functions on App module to Domain package
