+++ app/src/main/java/com/lucas/currencylist/domain/extensions/CurrencyExtensions.kt
@@ -1,4 +1,4 @@
-package com.lucas.currencylist.ui.extensions
+package com.lucas.currencylist.domain.extensions
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
