+++ app/src/main/java/com/lucas/currencylist/domain/extensions/ExchangePlatformExtensions.kt
@@ -1,4 +1,4 @@
-package com.lucas.currencylist.ui.extensions
+package com.lucas.currencylist.domain.extensions
 
 import androidx.annotation.DrawableRes
 import com.lucas.core.data.models.ExchangePlatformType
